# Improve performance of FromBytesOrNil