@cameracker 2025-03-26
@kohenkatz  Not sure either. The speed up for the invalid cases is rather substantial, but how often in the wild would this case actually be encountered and on the happy path of a consumer's workflow? The sub nanosecond time increase doesn't worry me that much. Willing to accept this trade, but will probably give this a couple of weeks to see if anybody pipes up and raises a concern.

@gaffneyc  would you mind adding a unit test to make sure we're at 100% coverage?