@gaffneyc 2025-03-27
I took a look at the code coverage change and since we're checking the only case where `FromBytes` can return an error there was no way to test the conditional which is what caused the drop in test coverage. My intention with the original change was to make the smallest change necessary as there may be other cases in the future where `FromBytes` could return an error (though I can't think of any).

That said, I've pushed an alternative implementation which pulls code from `FromBytes` and `UnmarshalBinary` into `FromBytesOrNil`. This avoids checking the length of the input twice as well as the overhead involved with calling multiple functions but duplicates the (minimal) logic for unmarshalling from a byte slice. I've also rerun the benchmarks on a less busy system and switched to benchstat for comparison.

By the way, the extra overhead of 0.65ns across 3 trillion calls would be [32 minutes and 30 seconds](https://www.wolframalpha.com/input?i2d=true&i=0.65ns+*+3+trillion) rather than 22 hours and 30 minutes.

I've updated the benchmarks in the main comment for the updated implementation. In the example with 3 trillion requests this would save [~2 hours and 15 minutes](https://www.wolframalpha.com/input?i2d=true&i=2.6965ns+*+3+trillion) of compute time per day for valid UUIDs and [~112 hours and 15 minutes](https://www.wolframalpha.com/input?i2d=true&i=134.6986ns+*+3+trillion) if all of those UUIDs were empty.