@dylan-bourque 2025-03-27
Another option here to avoid allocating errors that are just discarded would be to change `UnmarshalBinary()` to return a constant error rather than using `fmt.Errorf()`