@gaffneyc 2025-03-27
@dylan-bourque Yes, I rewrote the implementation pulling code from `FromBytes` and `UnmarshalBinary` into `FromBytesOrNil`. This was the only way without changing `UnmarshalBinary` to avoid checking the input's length twice and to get code coverage back to 100%.

> Another option here to avoid allocating errors that are just discarded would be to change UnmarshalBinary() to return a constant error rather than using fmt.Errorf()

Yeah, that could work as well though it would be a change in behavior. To keep the same behavior you could implement a custom type which would wrap the `ErrIncorrectByteLength`, store the invalid size, and respond with the full message when `Error()` is called. Not sure what impact either option would have without benchmarking it.

The original implementation was:
```go
func FromBytesOrNil(input []byte) UUID {
	if len(input) != Size {
		return Nil
	}

	uuid, err := FromBytes(input)
	if err != nil {
		return Nil
	}
	return uuid
}
```

The new implementation is:
```go
func FromBytesOrNil(input []byte) UUID {
	if len(input) != Size {
		return Nil
	}

	uuid := UUID{}
	copy(uuid[:], input)

	return uuid
}
```